### 2018-09-21 #969 @whedon
[PRE REVIEW]: perfectns: perfect dynamic and standard nested sampling for spherically symmetric likelihoods and priors
**Submitting author:** @ejhigson (<a href="http://orcid.org/0000-0001-8383-4614">Edward Higson</a>)
**Repository:** <a href="https://github.com/ejhigson/perfectns" target ="_blank">https://github.com/ejhigson/perfectns</a>
**Version:** v2.1.0
