### 2018-09-22 2c6e0da
making posterior mass pannel wider in logx plots
