### 2018-09-24 179be7e
minor edits to paper.md
