### 2018-09-24 c2c7ba8
minor paper edits
