### 2018-09-26 16eb670
minor paper updates
