### 2018-09-27 318ed3f
improving docs homepage
