### 2018-09-27 9b62265
added maths symbols to theory
