### 2018-09-27 #969 @mattpitkin
@ejhigson sorry, I won't be able to do this on a short timescale
