### 2018-09-28 #969 @adam-m-jcbs
> @mattpitkin @nespinoza @adam-m-jcbs would you be willing to review this submission? No worries if not, I can suggest alternative reviewers.
>
> Thank you very much for your help with this - it is much appreciated! I am keen to get the review started as I am about to leave my current position and want to wrap up my projects.

Thank you for considering me.  A good-faith review over a short timescale is currently not feasible for me, especially as the statistical theory underlying this project is not something I actively use in my current research and thus would require me to read up on the fundamental literature you cite.  In addition, it would be my first JOSS review and doing something for the first time usually takes me a while.

I hope next time I'll be in a better position to carry out a timely review.
