### 2018-09-29 #969 @arfon
@ziotom78, @ejhigson - see you over in #985 for the actual review.
