### 2018-10-01 #985 @ziotom78
I've read the paper; it looks ok, there are only one spelling error and one unclear sentence:

- Third paragraph: "such problems provides" should be "such problems provide"
- Same sentence: I do not fully understand the meaning of the clause "…and for use statistical research into nested sampling."
