### 2018-10-01 f76019d
typo fixes in paper
