### 2018-10-01 #985 @ejhigson
> Hi @ejhigson, don't worry, I should be able to complete the review in a few days.

Thank you so much for this! It is much appreciated.

> I've read the paper; it looks ok, there are only one spelling error and one unclear sentence:
>
> * Third paragraph: "such problems provides" should be "such problems provide"
> * Same sentence: I do not fully understand the meaning of the clause "…and for use statistical research into nested sampling."

Thank you for pointing these out! They are both typos. The sentence now reads "Furthermore, such problems provide a rich source of test cases for assessing the capacity of other software implementations to perform the nested sampling algorithm accurately, and for use in statistical research into nested sampling."
