### 2018-10-01 #985 @ziotom78
@ejhigson , I noticed that some docstrings are not correct:

- The docstring for [`priors.Uniform`](https://github.com/ejhigson/perfectns/blob/master/perfectns/priors.py#L32) and [`priors.Gaussian`](https://github.com/ejhigson/perfectns/blob/master/perfectns/priors.py#L69) are identical

- In the docstrings of these two classes, it might be worth expliciting what is the meaning of the parameter `prior_scale` passed to `__init__` (e.g., in the Gaussian prior it's just the sigma).
