### 2018-10-02 2a68a7c
added prior doc
