### 2018-10-03 f692c89
updating dyPolyChord in .bib following publication in JOSS
