+++ docs/paper/paper.bib
@@ -165,10 +165,14 @@ title = {{Dark Energy Survey Year 1 Results: Cosmological Constraints from Galax
 volume = {98},
 year = {2018}
 }
-@misc{Higson2018dypolychord,
+@article{Higson2018dypolychord,
 author = {Higson, Edward},
-doi = {10.5281/zenodo.1328175},
+doi = {10.21105/joss.00965},
+journal = {Journal of Open Source Software},
+number = {29},
+pages = {965},
 title = {{dyPolyChord: dynamic nested sampling with PolyChord}},
+volume = {3},
 year = {2018}
 }
 @article{zenodoperfectns,
