### 2018-10-03 858d43d
minor fix for res startin with lower case letter
