+++ docs/paper/paper.bib
@@ -134,7 +134,7 @@ doi = {10.21105/joss.00916},
 journal = {Journal of Open Source Software},
 number = {29},
 pages = {916},
-title = {nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations},
+title = {{nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations}},
 volume = {3},
 year = {2018}
 }
@@ -190,7 +190,7 @@ year = {2001}
 @misc{mpmath,
 key = {mpmath},
 author = {Fredrik Johansson and others},
-title = {mpmath: a Python library for arbitrary-precision floating-point arithmetic},
+title = {{mpmath: a Python library for arbitrary-precision floating-point arithmetic}},
 url = {http://mpmath.org/},
 year = {2013},
 }
