+++ docs/paper/paper.bib
@@ -217,7 +217,7 @@ pages = {90--95},
 title = {{Matplotlib: A 2D graphics environment}},
 volume = {9},
 year = {2007},
-doi = {https://doi.org/10.1109/MCSE.2007.55}
+doi = {10.1109/MCSE.2007.55}
 }
 @article{Keeton2011,
 author = {Keeton, Charles R.},
