### 2018-10-03 #985 @ejhigson
@arfon thank you very much for your help with this! I have archived the latest version at:

10.5281/zenodo.1443509
